% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos.R
\name{buscaProfundidade}
\alias{buscaProfundidade}
\title{buscaProfundidade}
\usage{
buscaProfundidade(x, vertice, visitados = c(), arestasOut = c(), matriz = NULL)
}
\arguments{
\item{matriz}{}
}
\description{
buscaProfundidade
}
