% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos.R
\name{distanciaCaminhoMinimo}
\alias{distanciaCaminhoMinimo}
\title{distanciaCaminhoMinimo}
\usage{
distanciaCaminhoMinimo(x, vertice, verticeDestino)
}
\arguments{
\item{verticeDestino}{}
}
\description{
distanciaCaminhoMinimo
}
